added authentication

+++ src/components/WorkoutDetails.js
@@ -1,13 +1,26 @@
 import {formatDistanceToNow} from 'date-fns/formatDistanceToNow';
+import { useAuthContext } from '../hooks/useAuthContext';
+import { useWorkoutsContext } from '../hooks/useWorkoutsContext';
 
 const WorkoutDetails = ({ workout }) => {
 
+    const { user } = useAuthContext();
+    const { dispatch } = useWorkoutsContext();
+
     const handleClick = async () => {
+        if(!user) {
+            return ;
+        }
         const response = await fetch('/api/workouts/' + workout._id, {
             method: 'DELETE',
+            headers: {
+                'Authorization': `Bearer ${user.token}`
+            }
         })
+        const json = await response.json();
+
         if(response.ok) {
-            console.log('deleted workout', response.json())
+            dispatch({type: 'DELETE_WORKOUT', payload: json})
         }
     }
     return(
